4_ray_adk_mcp/adk_mcp_app/requirements.txt
google_adk>=0.5.0
fastapi>=0.103.1
pydantic>=2.11.4,<3
litellm>=1.30.0
httpx>=0.27.0

# MCP dependencies
mcp==1.8.1
uvicorn>=0.23.2